Slightly revise framerate limiter and r_showfps 2 display
git-svn-id: https://svn.eduke32.com/eduke32@7304 1a8010ca-5511-0410-912e-c29ae57300e0

+++ source/duke3d/src/global.h
@@ -160,10 +160,10 @@ G_EXTERN projectile_t SpriteProjectile[MAXSPRITES];
 G_EXTERN sound_t g_sounds[MAXSOUNDS];
 G_EXTERN uint32_t everyothertime;
 G_EXTERN uint32_t g_moveThingsCount;
-G_EXTERN uint32_t g_gameUpdateTime;
-G_EXTERN uint32_t g_gameUpdateAndDrawTime;
+G_EXTERN double g_gameUpdateTime;
+G_EXTERN double g_gameUpdateAndDrawTime;
 #define GAMEUPDATEAVGTIMENUMSAMPLES 100
-extern float g_gameUpdateAvgTime;
+extern double g_gameUpdateAvgTime;
 
 #ifndef global_c_
 extern char CheatKeys[2];
